A few improvements
Remaining issue: if both player's and dealer's counts are equal to 20 and at the prompt you choose stay (as you should), the game loops around "hit" or "stay", because so does the dealer infinitely. 
A way around this is not to allow the dealer to choose if he hits or stays. Just hit him automatically :) 
Not sure what the game rules are exactly.